backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Parsery
cfgrib==0.9.14.0
netCDF4==1.7.1
h5py==3.11.0
xarray==2024.9.0
eccodes==1.7.1

# Geo — KLUCZOWE dla ODIM HDF5
pyproj==3.6.1
rasterio==1.3.11
shapely==2.0.6
fiona==1.9.6
geopandas==1.0.1
rio-tiler==6.7.0

# Utils
numpy==2.1.1
pandas==2.2.3
scipy==1.14.1
matplotlib==3.9.2
Pillow==10.4.0
python-dotenv==1.0.1
